% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm10.R
\docType{data}
\name{pm10}
\alias{pm10}
\title{Tibble con i dati di pm10 per 5 stazioni della regione PATRENTO}
\format{
Un tibble con 8 colonne e 14000 osservazioni
}
\usage{
pm10
}
\description{
Tibble con i dati di pm10 per 5 stazioni della regione PATRENTO
}
\keyword{datasets}
